Move to System.Diagnostics
