+++ src/libraries/System.Data.Common/tests/System/Data/RestrictedTypeHandlingTests.cs
@@ -3,6 +3,7 @@
 
 using System.Collections.Generic;
 using System.Data.SqlTypes;
+using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using System.Numerics;
