+++ src/libraries/System.Runtime/tests/System/Diagnostics/UnreachableExceptionTests.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Diagnostics;
 using Xunit;
 
 namespace System.Tests
