+++ src/tests/Interop/COM/ComWrappers/API/Program.cs
@@ -6,6 +6,7 @@ namespace ComWrappersTests
     using System;
     using System.Collections;
     using System.Collections.Generic;
+    using System.Diagnostics;
     using System.Runtime.CompilerServices;
     using System.Runtime.InteropServices;
