+++ src/tests/Interop/ObjectiveC/ObjectiveCMarshalAPI/Program.cs
@@ -5,6 +5,7 @@ namespace ObjectiveCMarshalAPI
 {
     using System;
     using System.Collections.Generic;
+    using System.Diagnostics;
     using System.Reflection;
     using System.Runtime.InteropServices;
     using System.Runtime.InteropServices.ObjectiveC;
